Services/ProfileService/requirements.txt
Flask
python-dotenv
Flask-SQLAlchemy
Flask-Migrate
pymysql
flask-marshmallow